# docs/swagger.yaml
definitions:
  dto.NewItemRequest:
    properties:
      description:
        type: string
      itemCode:
        type: string
      quantity:
        type: integer
    type: object
  dto.NewOrderRequest:
    properties:
      customerName:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.NewItemRequest'
        type: array
      orderedAt:
        type: string
    type: object
  models.Item:
    properties:
      createdAt:
        type: string
      description:
        type: string
      itemCode:
        type: string
      itemID:
        type: integer
      orderID:
        type: integer
      quantity:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Order:
    properties:
      createdAt:
        type: string
      customerName:
        type: string
      items:
        items:
          $ref: '#/definitions/models.Item'
        type: array
      orderID:
        type: integer
      orderedAt:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /orders:
    get:
      consumes:
      - application/json
      description: Get all order data
      operationId: get-all-order
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create an Order with Items
      operationId: create-order
      parameters:
      - description: Create order request body
        in: body
        name: Order
        required: true
        schema:
          $ref: '#/definitions/dto.NewOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Order'
      tags:
      - Orders
...
swagger: "2.0"